+++ RealmDemo/AppCustoms/SelectAttachmentController/AttachDocumentView/AttachDocumentCell.m
@@ -32,7 +32,7 @@
     [self.attachImage setImage:[[UIImage imageNamed:[NSString stringWithFormat:@"ic_%@",attachment.fileName.pathExtension]] imageWithRenderingMode:UIImageRenderingModeAlwaysTemplate] forState:UIControlStateNormal];
     self.attachName.text = attachment.fileName;
     self.attachSize.text = [NSString stringWithFormat:@"%.3fM",attachment.fileSize / 1024.f / 1024.f];
-    self.attachCreateDate.text = [NSString stringWithFormat:@"%d-%02ld-%02ld %02ld:%02ld",attachment.fileCreateDate.year,attachment.fileCreateDate.month,attachment.fileCreateDate.day,attachment.fileCreateDate.hour,attachment.fileCreateDate.minute];
+    self.attachCreateDate.text = [NSString stringWithFormat:@"%ld-%02ld-%02ld %02ld:%02ld",(long)attachment.fileCreateDate.year,attachment.fileCreateDate.month,attachment.fileCreateDate.day,attachment.fileCreateDate.hour,attachment.fileCreateDate.minute];
 }
 
 @end
