+++ RealmDemo/ViewControllers/BusinessController/TabBarController/HomeListController/TaskListController/TaskCreateController/TaskCreateCell/TaskRemindCell.m
@@ -21,7 +21,7 @@
 }
 - (void)dataDidChange {
     NSDate *date = self.data;
-    self.dateLabel.text = [NSString stringWithFormat:@"%d-%02ld-%02ld %02ld:%02ld",date.year,date.month,date.day,date.hour,date.minute];;
+    self.dateLabel.text = [NSString stringWithFormat:@"%ld-%02ld-%02ld %02ld:%02ld",(long)date.year,date.month,date.day,date.hour,date.minute];;
 }
 
 - (IBAction)deleClicked:(id)sender {
